corrigiendo fitltros, se implementa validarQR completo

+++ src/main/java/com/gym/gym_ver2/domain/model/dto/RegisterRequestDTO.java
@@ -26,8 +26,8 @@ public class RegisterRequestDTO {
     Date fechaNacimiento;
     String sexo;
 
-    BigDecimal estatura;
-    BigDecimal peso;
+    Double estatura;
+    Double peso;
     Integer ficha;
     String jornada;
     Integer horasAcumuladas;
